Move interfaces and types to dedicated folders
