+++ components/AuthorCardList.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import AuthorCollection from '../util/AuthorCollection'
+import AuthorCollection from '../interfaces/AuthorCollection'
 
 import AuthorCard from './AuthorCard'
