+++ interfaces/AuthorCollection.ts
