+++ interfaces/MarkdownFile.ts
@@ -6,5 +6,3 @@ export default interface MarkdownFile {
     frontMatter: FrontMatter
     body: string
 }
-
-export type MarkdownFileWithUrl = MarkdownFile & { url: string }
