+++ interfaces/RecordCollection.ts
