+++ interfaces/RecordCollectionDefinition.ts
