+++ pages/index.tsx
@@ -4,11 +4,9 @@ import React from 'react'
 import ContentCardList from '../components/ContentCardList'
 import Header from '../components/Header'
 import PageLayout from '../components/PageLayout'
+import { MarkdownFileWithUrl } from '../types/MarkdownFileWithUrl'
 import loadAllRecords from '../util/loadAllRecords'
-import MarkdownFile from '../util/MarkdownFile'
 import omitUndefinedFields from '../util/omitUndefinedFields'
-
-export type MarkdownFileWithUrl = MarkdownFile & { url: string }
 
 interface Props {
     posts: MarkdownFileWithUrl[]
