+++ types/MarkdownFileWithUrl.ts
@@ -1,0 +1,3 @@
+import MarkdownFile from '../interfaces/MarkdownFile'
+
+export type MarkdownFileWithUrl = MarkdownFile & { url: string }
