+++ util/collectTags.ts
@@ -1,6 +1,6 @@
 import uniq from 'lodash/uniq'
 
-import MarkdownFile from './MarkdownFile'
+import MarkdownFile from '../interfaces/MarkdownFile'
 
 export default function collectTags(markdownFiles: MarkdownFile[]): string[] {
     const tags = markdownFiles.flatMap(file => file.frontMatter.tags)
